Overhaul makefiles and header layout.
This commit replaces the Makefile build system in wxLua with one which
doesn't make use of recursive make and allows for more flexibility when
selecting the wxWidgets/Lua libraries to build/link against.

The end goal of this work is to make it so that portions of wxLua can
be compiled into other Make-based projects by including the Makefile
and simply depending on the appropriate targets.

The C++ #include directives are reworked so that relative include paths
are *actually* relative (where appropriate) rather than depending on
the compiler include path including practically every directory in the
wxLua source tree.

The public headers have been moved to a dedicated include directory.